chatbot-salud-gq/requirements.txt
flask==3.1.0
requests==2.32.3
openai==1.58.1
python-dotenv==1.0.1
gunicorn==23.0.0
